fix/backend dir refactor
