+++ backend/src/Chat/index.ts
@@ -1,0 +1,3 @@
+import { ChatModule } from './chat.module';
+
+export { ChatModule };
