+++ backend/src/chatMuteUser/index.ts
@@ -1,0 +1,3 @@
+import { ChatMuteUserModule } from './chatMuteUser.module';
+
+export { ChatMuteUserModule };
