+++ backend/src/healthCheck/healthCheck.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { HealthCheck } from './healthCheck.entity';
+import { HealthCheck } from '../entity/healthCheck.entity';
 import { HealthCheckService } from './healthCheck.service';
 import { HealthCheckController } from './healthCheck.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
