+++ backend/src/healthCheck/index.ts
@@ -1,0 +1,3 @@
+import { HealthCheckModule } from './healthCheck.module';
+
+export { HealthCheckModule };
