+++ backend/typings/chat.d.ts
@@ -1,0 +1,6 @@
+interface messageEventType {
+  key: number;
+  name: string;
+  room: string;
+  msg: string;
+}
